feat: add service functions and API routes for signup and login with JWT authentication

+++ backend/src/services/posts.js
@@ -1,9 +1,7 @@
-import mongoose from 'mongoose'
-
 import { Post } from '../db/models/post.js'
 
-export async function createPost({ title, author, contents, tags }) {
-  const post = new Post({ title, author, contents, tags })
+export async function createPost(userId, { title, contents, tags }) {
+  const post = new Post({ title, author: userId, contents, tags })
   return await post.save()
 }
 
@@ -19,9 +17,9 @@ export async function listAllPosts(options) {
 }
 
 export async function listPostsByAuthor(author, options) {
-  if (!mongoose.Types.ObjectId.isValid(author)) {
-    throw new Error('Invalid author ID')
-  }
+  // if (!mongoose.Types.ObjectId.isValid(author)) {
+  //   throw new Error('Invalid author ID')
+  // }
 
   return await listPosts({ author }, options)
 }
@@ -34,18 +32,19 @@ export async function getPostById(postId) {
   return await Post.findById(postId)
 }
 
-export async function updatePost(postId, { title, author, contents, tags }) {
-  if (!mongoose.Types.ObjectId.isValid(author)) {
-    throw new Error('Invalid author ID')
-  }
+// TODO: update function to not allow the author to be changed
+export async function updatePost(userId, postId, { title, contents, tags }) {
+  // if (!mongoose.Types.ObjectId.isValid(author)) {
+  //   throw new Error('Invalid author ID')
+  // }
 
   return await Post.findOneAndUpdate(
-    { _id: postId },
-    { $set: { title, author, contents, tags } },
+    { _id: postId, author: userId },
+    { $set: { title, contents, tags } },
     { new: true },
   )
 }
 
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId })
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
